Keep the server's integration graph out of the web typecheck
